refactor(webdriverio): extract shared test utilities into testUtils.js
Move getFreePort, connectToChromeDriver, and BDM env loading into a
shared testUtils.js so both esmTest.mjs and axe-webdriverio.spec.ts
can import them rather than duplicating the logic.
