+++ packages/webdriverio/test/esmTest.mjs
@@ -6,9 +6,7 @@ import { pathToFileURL } from 'url';
 import { join } from 'path';
 import { fixturesPath } from 'axe-test-fixtures';
 import { spawn } from 'child_process';
-import net from 'net';
-import os from 'os';
-import { readFileSync } from 'fs';
+import { getFreePort, connectToChromeDriver, loadBdmEnv } from './testUtils.js';
 
 assert(typeof defaultExport === 'function', 'default export is not a function');
 assert(typeof AxeBuilder === 'function', 'named export is not a function');
@@ -17,57 +15,7 @@ assert(
   'default and named export are not the same'
 );
 
-// Load browser-driver-manager env vars (same approach as axe-webdriverio.spec.ts)
-const BDM_ENV_PATH = join(os.homedir(), '.browser-driver-manager', '.env');
-try {
-  const envContent = readFileSync(BDM_ENV_PATH, 'utf-8');
-  for (const line of envContent.split('\n')) {
-    const match = line.match(/^([^=\s]+)=(.*)$/);
-    if (match) {
-      // Strip surrounding quotes that some .env writers add
-      const value = match[2].replace(/^(['"])(.*)\1$/, '$2');
-      process.env[match[1]] = value;
-    }
-  }
-} catch {
-  // .env file not found; env vars may already be set in the environment
-}
-
-const connectToChromeDriver = port => {
-  return new Promise((resolve, reject) => {
-    const timer = setTimeout(() => {
-      socket.destroy();
-      reject(new Error('Unable to connect to ChromeDriver'));
-    }, 1000);
-    const socket = net.createConnection({ host: 'localhost', port }, () => {
-      clearTimeout(timer);
-      socket.destroy();
-      resolve();
-    });
-    socket.once('error', err => {
-      clearTimeout(timer);
-      socket.destroy();
-      reject(err);
-    });
-  });
-};
-
-const getFreePort = () => {
-  return new Promise((resolve, reject) => {
-    const server = net.createServer();
-    server.listen(0, '127.0.0.1', () => {
-      const { port } = server.address();
-      server.close(err => {
-        if (err) {
-          reject(err);
-        } else {
-          resolve(port);
-        }
-      });
-    });
-    server.once('error', reject);
-  });
-};
+loadBdmEnv();
 
 async function integrationTest() {
   const port = await getFreePort();
